--- src/main/java/likelion15/mutsa/dto/User.java
@@ -1,13 +1,0 @@
-package likelion15.mutsa.dto;
-
-import lombok.Data;
-
-@Data
-public class User {
-    private Long id;
-    private String username;
-    private String email;
-    private String password;
-    private String phoneNumber;
-
-}
