+++ src/main/java/likelion15/mutsa/repository/UserRepos.java
@@ -3,10 +3,7 @@ package likelion15.mutsa.repository;
 import likelion15.mutsa.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User,Long> {
+public interface UserRepos extends JpaRepository<User,Long> {
 
     User findByEmail(String email);
 }
